Flatten parser directory
We had a 'doxygen' subfolder but we're never going to support another
source format as this project doesn't have that level of scope so the
extra level of namespacing is not required.

We also fix pep-8 issues though I should have only updated the imports
and done the other changes in another commit.